Updating Smart Contracts premint values, creating deployContract and createDeploymentModule

+++ backend/contracts/Tokgen/ERC20/Ownable/1-Single/TokgenERC20FlashMintOwnable.sol
@@ -10,7 +10,7 @@ contract TokgenERC20FlashMintOwnable is ERC20, Ownable {
         address initialOwner,
         string memory name,
         string memory symbol,
-        uint8 premint
+        uint256 premint
     ) ERC20(name, symbol) Ownable(initialOwner) {
         _mint(msg.sender, premint * 10 ** decimals());
     }
