+++ backend/contracts/Tokgen/ERC20/Ownable/1-Single/TokgenERC20PausableOwnable.sol
@@ -11,7 +11,7 @@ contract TokgenERC20PausableOwnable is ERC20, ERC20Pausable, Ownable {
         address initialOwner,
         string memory name,
         string memory symbol,
-        uint8 premint
+        uint256 premint
     ) ERC20(name, symbol) Ownable(initialOwner) {
         _mint(msg.sender, premint * 10 ** decimals());
     }
