+++ backend/contracts/Tokgen/ERC20/Ownable/1-Single/TokgenERC20PermitOwnable.sol
@@ -11,7 +11,7 @@ contract TokgenERC20PermitOwnable is ERC20, ERC20Permit, Ownable {
         address initialOwner,
         string memory name,
         string memory symbol,
-        uint8 premint
+        uint256 premint
     ) ERC20(name, symbol) ERC20Permit(name) Ownable(initialOwner) {
         _mint(msg.sender, premint * 10 ** decimals());
     }
