+++ backend/contracts/Tokgen/ERC20/Ownable/2-Double/TokgenERC20PausablePermitOwnable.sol
@@ -12,7 +12,7 @@ contract TokgenERC20PausablePermitOwnable is ERC20, ERC20Pausable, Ownable, ERC2
         address initialOwner,
         string memory name,
         string memory symbol,
-        uint8 premint
+        uint256 premint
     ) ERC20(name, symbol) Ownable(initialOwner) ERC20Permit(name) {
         _mint(msg.sender, premint * 10 ** decimals());
     }
