+++ backend/contracts/Tokgen/ERC20/Ownable/4-Quadruple/TokgenERC20MintableBurnablePausableFlashMintOwnable.sol
@@ -19,7 +19,7 @@ contract TokgenERC20MintableBurnablePausableFlashMintOwnable is
         address initialOwner,
         string memory name,
         string memory symbol,
-        uint8 premint
+        uint256 premint
     ) ERC20(name, symbol) Ownable(initialOwner) {
         _mint(msg.sender, premint * 10 ** decimals());
     }
